// entidad-web/entidad-web/src/app/@data/model/sede.ts
export class Sede {
    sedeId: number;
    nombreSede: string;
    direccionId: number;
    direccion: string;
    departamentoId: number;
    departamento: string;
    provinciaId: number;
    provincia: string;
    distritoId: number;
    distrito: string;
    estado: string;
    estadoId: string;

    representante: string;
    telefono: string;
    anexo: string;

    settings?: any;
}